+++ src/main/java/org/lin/service/UserServiceImp.java
@@ -6,6 +6,7 @@ import org.lin.mapper.UserMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import javax.servlet.http.HttpServletRequest;
 import java.util.List;
 import java.util.UUID;
 
@@ -14,6 +15,8 @@ public class UserServiceImp implements UserService{
 
   @Autowired
   private UserMapper userMapper;
+  @Autowired
+  private TokenMapper tokenMapper;
 
   @Override
   public Integer login(User user) {
@@ -26,18 +29,36 @@ public class UserServiceImp implements UserService{
   }
 
   @Override
-  public int addUser(User user) {
-    return userMapper.addUser(user);
+  public int addUser(User user, String token) {
+    Integer power = tokenMapper.realGetPowerByToken(token);
+    if(power == 0) {
+      return userMapper.addUser(user);
+
+    }else {
+      return 0;
+    }
   }
 
   @Override
-  public int deleteUser(int id) {
-    return userMapper.deleteUser(id);
+  public int deleteUser(int id, String token) {
+
+    Integer power = tokenMapper.realGetPowerByToken(token);
+    if(power == 0) {
+      return userMapper.deleteUser(id);
+    }else {
+      return 0;
+    }
   }
 
   @Override
-  public int editUser(User user) {
-    return userMapper.editUser(user);
+  public int editUser(User user, String token) {
+
+    Integer power = tokenMapper.realGetPowerByToken(token);
+    if(power == 0) {
+      return userMapper.editUser(user);
+    }else {
+      return 0;
+    }
   }
 
   @Override
